cursor/claude4 refactor of TokenList
