+++ wormhole-connect/src/views/v2/Bridge/AssetPicker/TokenList.tsx
@@ -1,33 +1,18 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import React, { useEffect, useMemo } from 'react';
 import { Box, Card, CardContent, Skeleton, useTheme } from '@mui/material';
 import CircularProgress from '@mui/material/CircularProgress';
 import ListItemButton from '@mui/material/ListItemButton';
 import Typography from '@mui/material/Typography';
 import { makeStyles } from 'tss-react/mui';
-import {
-  circle,
-  isNative,
-  amount as sdkAmount,
-  toNative,
-} from '@wormhole-foundation/sdk';
+import { toNative } from '@wormhole-foundation/sdk';
 
 import useGetTokenBalances from 'hooks/useGetTokenBalances';
 import type { ChainConfig } from 'config/types';
-import {
-  isTokenTuple,
-  isSameToken,
-  Token,
-  tokenIdFromTuple,
-  tokenKey,
-} from 'config/tokens';
+import { Token } from 'config/tokens';
 import type { WalletData } from 'store/wallet';
 import SearchableList from 'views/v2/Bridge/AssetPicker/SearchableList';
 import TokenItem from 'views/v2/Bridge/AssetPicker/TokenItem';
-import {
-  calculateUSDPrice,
-  calculateUSDPriceRaw,
-  isFrankensteinToken,
-} from 'utils';
+import { calculateUSDPrice } from 'utils';
 import config from 'config';
 import { useTokens } from 'contexts/TokensContext';
 
@@ -57,14 +42,15 @@ const useStyles = makeStyles()((theme: any) => ({
 }));
 
 type Props = {
-  tokenList?: Array<Token>;
+  tokenList: Array<Token>; // Remove optional since we'll pass sorted tokens
   isFetching?: boolean;
   selectedChainConfig: ChainConfig;
   selectedToken?: Token;
   sourceToken?: Token;
   wallet: WalletData;
+  searchQuery: string;
+  onSearchQueryChange: (query: string) => void;
   onSelectToken: (key: Token) => void;
-  isSource: boolean;
 };
 
 const TokenList = (props: Props) => {
@@ -74,7 +60,7 @@ const TokenList = (props: Props) => {
 
   const { getOrFetchToken, isFetchingToken, getTokenPrice } = useTokens();
 
-  const [searchQuery, setSearchQuery] = useState('');
+  // Search query is now managed by parent component
 
   const { isFetching: isFetchingTokenBalances, balances } = useGetTokenBalances(
     props.wallet,
@@ -88,12 +74,12 @@ const TokenList = (props: Props) => {
     // If not, try to find it.
     if (tokenPastingIsEnabled) {
       try {
-        if (searchQuery !== '') {
+        if (props.searchQuery !== '') {
           const chain = props.selectedChainConfig.sdkName;
-          const address = toNative(chain, searchQuery);
+          const address = toNative(chain, props.searchQuery);
 
           if (address) {
-            const existing = config.tokens.get(chain, searchQuery);
+            const existing = config.tokens.get(chain, props.searchQuery);
             if (!existing) {
               getOrFetchToken({ chain, address });
             }
@@ -105,177 +91,10 @@ const TokenList = (props: Props) => {
     }
     // Run the side-effect only when search query or chain changes
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [searchQuery, props.selectedChainConfig.sdkName]);
+  }, [props.searchQuery, props.selectedChainConfig.sdkName]);
 
-  // Returns a score for a given token used when sorting destination tokens
-  const tokenPreferenceScore = (token: Token) => {
-    // Currently selected token should be shown first
-    if (props.selectedToken && isSameToken(props.selectedToken, token)) {
-      return 4;
-    }
-    // Native gas tokens are next
-    if (isNative(token.addressString)) {
-      return 3;
-    }
-    // USDC preferred next
-    const usdc = circle.usdcContract.get(config.network, token.chain);
-    if (usdc && token.addressString === usdc) {
-      return 2;
-    }
-    // Finally, prefer native non-wrapped tokens over wrapped ones
-    if (!token.isTokenBridgeWrappedToken) {
-      return 1;
-    }
-    // The rest is all the same as far as preference
-    return 0;
-  };
-
-  // TODO this entire thing should be moved outside of this TokenList component. The component is doing way too much.
-  const sortedTokens = useMemo(() => {
-    if (!props.tokenList) return [];
-
-    const unsortedTokens = props.tokenList;
-
-    // Apply search input - find tokens with exact match of address, or partial match of symbol
-    if (searchQuery) {
-      let searchResults: Token[] = [];
-      const byAddress = config.tokens.get(
-        props.selectedChainConfig.sdkName,
-        searchQuery,
-      );
-      if (byAddress) {
-        searchResults.push(byAddress);
-      }
-
-      const queryResults = config.tokens
-        .queryBySymbol(props.selectedChainConfig.sdkName, searchQuery)
-        .filter(
-          (t: Token) =>
-            !isFrankensteinToken(t, props.selectedChainConfig.sdkName),
-        );
-
-      if (queryResults.length > 0) {
-        searchResults = searchResults.concat(queryResults);
-      }
-
-      for (const result of searchResults) {
-        if (
-          !props.tokenList.find((existing) => isSameToken(result, existing))
-        ) {
-          unsortedTokens.push(result);
-        }
-      }
-    }
-
-    const usdBalance = (token: Token): number => {
-      const balance = balances[tokenKey(token)];
-      if (!balance || !balance.balance) {
-        return 0;
-      }
-      return calculateUSDPriceRaw(getTokenPrice, balance.balance, token) ?? 0;
-    };
-
-    let sorted = unsortedTokens.sort((a, b) => {
-      const scoreA = tokenPreferenceScore(a);
-      const scoreB = tokenPreferenceScore(b);
-      if (scoreA > scoreB) return -1;
-      if (scoreB > scoreA) return 1;
-
-      const balanceA = usdBalance(a);
-      const balanceB = usdBalance(b);
-      if (balanceA !== balanceB) {
-        return balanceB - balanceA;
-      } else {
-        // If equal scores and USD balance, compare by symbol
-        return a.symbol.localeCompare(b.symbol);
-      }
-    });
-
-    if (config.tokenWhitelist && config.tokenWhitelist.length > 0) {
-      // If integrator has specified a token whitelist, filter the token list by this whitelist.
-      //
-      // The logic behind how this works is a little complicated. The whitelist is an array of (string | TokenTuple).
-      // The strings can be symbols like "USDC", which lets the integrator easily whitelist tokens across all supported chains.
-      //
-      // The way we handle symbols is that for each chain:
-      // 1. If there is a single native token with that symbol, we simply use that
-      // 2. If there is NO native token with that symbol but there is a wrapped token, we show that.
-      // 3. If there are somehow multiple wrapped tokens with that symbol, which all passed through the isFrankensteinToken check above,
-      //    we include them all but log a warning to the console for the integrator's benefit.
-      //
-      // The integrator can also specify exact tokens using TokenTuples like ["Solana", "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"].
-
-      const filteredTokens: Set<string> = new Set();
-      const desiredSymbols: string[] = [];
-
-      for (const item of config.tokenWhitelist) {
-        if (typeof item === 'string') {
-          // Treated as a symbol
-          desiredSymbols.push(item);
-        } else if (isTokenTuple(item)) {
-          const tokenId = tokenIdFromTuple(item);
-          if (tokenId.chain === props.selectedChainConfig.sdkName) {
-            filteredTokens.add(tokenId.address.toString());
-          }
-        }
-      }
-
-      for (const symbol of desiredSymbols) {
-        let foundNative = false;
-        const wrapped: Token[] = [];
-
-        for (const token of sorted) {
-          if (token.symbol === symbol) {
-            if (!token.isTokenBridgeWrappedToken) {
-              filteredTokens.add(token.address.toString());
-              foundNative = true;
-            } else {
-              wrapped.push(token);
-            }
-          }
-        }
-
-        if (!foundNative && wrapped.length > 0) {
-          for (const { address } of wrapped) {
-            filteredTokens.add(address.toString());
-          }
-
-          if (wrapped.length > 1) {
-            console.warn(
-              `Ambiguous token whitelist item "${symbol}"; found ${wrapped.length} matching wrapped tokens.`,
-            );
-          }
-        }
-      }
-
-      sorted = sorted.filter((token) =>
-        filteredTokens.has(token.address.toString()),
-      );
-    }
-
-    if (config.isTokenSupportedHandler) {
-      // The last step is to filter the tokens by the integrator's token support handler
-      sorted = sorted.filter(config.isTokenSupportedHandler);
-    }
-
-    if (props.isSource && props.wallet.address) {
-      sorted = sorted.filter((t) => {
-        const bal = balances[tokenKey(t)]?.balance;
-        return bal && sdkAmount.units(bal) > 0;
-      });
-    }
-
-    return sorted;
-  }, [
-    props.selectedChainConfig.sdkName,
-    props.selectedToken,
-    props.tokenList,
-    props.sourceToken,
-    props.isSource,
-    props.wallet?.address,
-    balances,
-    searchQuery,
-  ]);
+  // Token list is now pre-sorted and passed from parent
+  const sortedTokens = props.tokenList;
 
   const noTokensMessage = useMemo(
     () => (
@@ -319,7 +138,7 @@ const TokenList = (props: Props) => {
       }
       items={sortedTokens}
       onQueryChange={(query) => {
-        setSearchQuery(query);
+        props.onSearchQueryChange(query);
       }}
       filterFn={(token, query) => {
         if (query.length === 0) return true;
@@ -352,11 +171,8 @@ const TokenList = (props: Props) => {
         const price = balance
           ? calculateUSDPrice(getTokenPrice, balance, token)
           : null;
-        const disabled =
-          props.isSource &&
-          !!props.wallet?.address &&
-          !!balances &&
-          (!balance || sdkAmount.units(balance) === 0n);
+        // Disabled logic is now handled by the filtering hooks
+        const disabled = false;
 
         return (
           <TokenItem
