+++ wormhole-connect/src/views/v2/Bridge/AssetPicker/index.tsx
@@ -25,6 +25,9 @@ import TokenList from './TokenList';
 import AssetBadge from 'components/AssetBadge';
 import { Token } from 'config/tokens';
 import { joinClass } from 'utils/style';
+import { useSourceTokenList } from 'hooks/useSourceTokenList';
+import { useDestTokenList } from 'hooks/useDestTokenList';
+import useGetTokenBalances from 'hooks/useGetTokenBalances';
 
 const useStyles = makeStyles()((theme: any) => ({
   inputArea: {
@@ -105,7 +108,37 @@ const AssetPicker = (props: Props) => {
   const mobile = useMediaQuery(theme.breakpoints.down('sm'));
   const [showChainSearch, setShowChainSearch] = useState(false);
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+  const [searchQuery, setSearchQuery] = useState('');
   const { classes } = useStyles();
+
+  // Get token balances for filtering
+  const { balances } = useGetTokenBalances(
+    props.wallet,
+    props.chain,
+    props.tokenList || [],
+  );
+
+  // Use the appropriate hook based on isSource
+  const sourceTokens = useSourceTokenList({
+    tokenList: props.tokenList || [],
+    searchQuery,
+    selectedChainConfig: props.chain ? config.chains[props.chain] : ({} as any),
+    selectedToken: props.token,
+    wallet: props.wallet,
+    balances,
+  });
+
+  const destTokens = useDestTokenList({
+    tokenList: props.tokenList || [],
+    searchQuery,
+    selectedChainConfig: props.chain ? config.chains[props.chain] : ({} as any),
+    selectedToken: props.token,
+    sourceToken: props.sourceToken,
+    wallet: props.wallet,
+    balances,
+  });
+
+  const sortedTokens = props.isSource ? sourceTokens : destTokens;
 
   const popupState = usePopupState({
     variant: 'popover',
@@ -261,17 +294,18 @@ const AssetPicker = (props: Props) => {
           />
           {!showChainSearch && chainConfig && (
             <TokenList
-              tokenList={props.tokenList}
+              tokenList={sortedTokens}
               isFetching={props.isFetching}
               selectedChainConfig={chainConfig}
               selectedToken={props.token}
               sourceToken={props.sourceToken}
               wallet={props.wallet}
+              searchQuery={searchQuery}
+              onSearchQueryChange={setSearchQuery}
               onSelectToken={(key: Token) => {
                 props.setToken(key);
                 setIsDrawerOpen(false);
               }}
-              isSource={props.isSource}
             />
           )}
         </SwipeableDrawer>
@@ -310,17 +344,18 @@ const AssetPicker = (props: Props) => {
           />
           {!showChainSearch && chainConfig && (
             <TokenList
-              tokenList={props.tokenList}
+              tokenList={sortedTokens}
               isFetching={props.isFetching}
               selectedChainConfig={chainConfig}
               selectedToken={props.token}
               sourceToken={props.sourceToken}
               wallet={props.wallet}
+              searchQuery={searchQuery}
+              onSearchQueryChange={setSearchQuery}
               onSelectToken={(key: Token) => {
                 props.setToken(key);
                 popupState.close();
               }}
-              isSource={props.isSource}
             />
           )}
         </Popover>
